requirements.txt
python-gitlab==1.8.0
